+++ EKIHIRE/EkiHire.Core/Domain/Entities/Auditing/AuditedEntity.cs
@@ -8,7 +8,7 @@ namespace EkiHire.Core.Domain.Entities.Auditing
     /// A shortcut of <see cref="AuditedEntity{TPrimaryKey}"/> for most used primary key type (<see cref="int"/>).
     /// </summary>
     [Serializable]
-    public abstract class AuditedEntity : AuditedEntity<int>, IEntity
+    public abstract class AuditedEntity : AuditedEntity<long>, IEntity
     {
 
     }
