+++ EKIHIRE/EkiHire.Data/Repository/IRepository.cs
@@ -20,7 +20,7 @@ namespace EkiHire.Data.Repository
     /// A shortcut of <see cref="IRepository{TEntity,TPrimaryKey}"/> for most used primary key type (<see cref="int"/>).
     /// </summary>
     /// <typeparam name="TEntity">Entity type</typeparam>
-    public interface IRepository<TEntity> : IRepository<TEntity, int> where TEntity : class, IEntity<int>
+    public interface IRepository<TEntity> : IRepository<TEntity, long> where TEntity : class, IEntity<long>
     {
     }
